// projeto/web/angular_p2/src/api/generated/models/updateUserDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Pizzaria API
 * API para gerenciamento de pedidos e clientes.
 * OpenAPI spec version: 1.0
 */

export interface UpdateUserDto {
  /**
   * @minLength 3
   * @maxLength 255
   * @nullable
   */
  name?: string | null;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  roleId?: number | string;
}